Create README.md
We have  completely implemented Microsoft Power BI project end to end. Right from requirement gathering to publishing the report. We have imported the data from  CSV files, and then cleaned and transformed the data as required. After that we have applied various DAX and calculated column calculation to compute our requirement. After that we have focused on UI enhancement to make it more readable. We have also converted the time into the nearest hour so that we can analyse the hourly orders received by the restaurant. This metric is one of the most important KPIs for this analysis.